pyproject.toml
[project]
name = "agent-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.19.0",
    "litellm>=1.80.5",
    "requests>=2.31.0",
    "geopy>=2.4.0",
    "wikipedia-api>=0.6.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["multi_tool_agent", "city_info_expert"]
